src/requirements.txt
pillow
--prefer-binary